Move the netlink dependency out of util
Code depending on github.com/vishvananda/netlink can't be built for
Windows, so we need to ensure that any package which ends up used by
subctl doesn't pull it in.
